fix: :bug: Use SDL backend for Gamescope-Steam
Not specifying backend, or letting Gamescope determine the backend (likely DRM), makes Steam unusable with Performance Overlay/mangoapp. This issue is solved by manually specifying backend to SDL or wayland. However, this causes USB keyboard input to not be registered in Steam with Gamescope. For the time being, let's monitor SDL backend behavior.